+++ src/renderer/src/store/thunks/index.ts
@@ -5,6 +5,7 @@ export * from '#/renderer/src/store/thunks/modals';
 export * from '#/renderer/src/store/thunks/sync';
 export * from '#/renderer/src/store/thunks/aiChat';
 export * from '#/renderer/src/store/thunks/collectionRunner';
+export * from '#/renderer/src/store/thunks/settings';
 
 import type { AppDispatch } from '#/renderer/src/store/redux';
 import { setGeneralSettingsState } from '#/renderer/src/store/slices/settingsSlice';